controller, service, domain 생성
