+++ src/main/java/com/gdsc/game/dto/Action.java
@@ -1,4 +1,4 @@
-package com.gdsc.game;
+package com.gdsc.game.dto;
 
 
 interface Action {
