+++ src/main/java/com/gdsc/game/dto/DefenseAction.java
@@ -1,4 +1,4 @@
-package com.gdsc.game;
+package com.gdsc.game.dto;
 
 import java.util.Random;
